Generic SyncDomainStore
Refactor persistent sync to use ushort wire IDs and reusable domain types. Introduces SyncClientDomain and SyncDomainStore generic bases, plus PersistentSyncDomainMetadata attributes and naming helpers to infer domain names and known wire IDs. Persistent sync messages now carry DomainWireId (ushort) and message size calculations updated; PersistentSyncDomainCatalog/Registrar support lookups by name/wireId and a RegisterCurrent helper that uses class attributes. Server/client domain classes updated to use the new generic stores/clients and attribute-driven scenario/scalar configuration; default payload creation/serialization logic consolidated.

+++ LmpCommon/PersistentSync/PersistentSyncDomainKey.cs
@@ -7,7 +7,7 @@ namespace LmpCommon.PersistentSync
     /// </summary>
     public struct PersistentSyncDomainKey : IEquatable<PersistentSyncDomainKey>
     {
-        public PersistentSyncDomainKey(string name, byte wireId)
+        public PersistentSyncDomainKey(string name, ushort wireId)
         {
             if (string.IsNullOrEmpty(name))
             {
@@ -19,7 +19,7 @@ namespace LmpCommon.PersistentSync
         }
 
         public string Name { get; }
-        public byte WireId { get; }
+        public ushort WireId { get; }
         public PersistentSyncDomainId LegacyId => (PersistentSyncDomainId)WireId;
 
         public bool Equals(PersistentSyncDomainKey other)
@@ -51,7 +51,7 @@ namespace LmpCommon.PersistentSync
     /// </summary>
     public static class PersistentSyncDomain
     {
-        public static PersistentSyncDomainKey Define(string name, byte wireId)
+        public static PersistentSyncDomainKey Define(string name, ushort wireId)
         {
             return new PersistentSyncDomainKey(name, wireId);
         }
